+++ gd-core/src/components/ol/index.jsx
@@ -1,1 +1,2 @@
+export BaseControlPortal from './BaseControlPortal';
 export Map, { MapContext } from './Map';