+++ gd-gltg/CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GLGVO-640](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-640)
 - Added overall nitrate load text for IL drainage
   [GLGVO-641](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-641)
+- Added layer legend to the map
+  [GLGVO-648](https://opensource.ncsa.illinois.edu/jira/browse/GLGVO-648)
 
 ### Changed
 - Create Intervals as a Reusable Component